adapt debouncer tests with the changed codebase
