+++ crates/gitbutler-watcher/tests/fixtures/emit_needs_rescan_event.hjson
@@ -42,7 +42,6 @@
         cache: {
             /watch/file-a: 1
             /watch/file-b: 2
-            /watch/file-c: 3
         }
         events: {
             short: []
